+++ tests/test_utilities.py
@@ -101,6 +101,19 @@ class TestUtilities(unittest.TestCase):
         h = pd.DataFrame()
         self.assertFalse(almost_equal(f, h))
 
+    def test_is_not_null(self):
+        f = None
+        c = 'foo'
+        self.assertFalse(is_not_null(f, c))
+
+        f = pd.DataFrame(columns=['bar', c])
+        self.assertFalse(is_not_null(f, c))
+
+        f = pd.DataFrame([[1, np.nan]], columns=['bar', c])
+        self.assertFalse(is_not_null(f, c))
+
+        f = pd.DataFrame([[1, np.nan], [2, 2]], columns=['bar', c])
+        self.assertTrue(is_not_null(f, c))
 
 if __name__ == '__main__':
     unittest.main()